fix: resolve toggle bug
When toggling state (mute, solo, ...) it sometimes didn't work at the first time.
This was due to shareReplay() which uses refCount() internally and therefore, doesn't connect to the source before the first subscriber. However, if there is no subscriber, the state info is missing and the toggling (based on current state) does not work.
fix #6