refactor: store state as flat object instead of nested structure
The mixer state was previously stored as a big nested object that was built from the incoming messages. However, we never needed this raw state object directly. Instead, we just extracted parts of it which all resolve to primitive values. It was therefore not necessary to populate the big object at all.
The state is now stored in a flatter structure. The keys are the unchanged paths from the mixer messages.

BREAKING CHANGE
The `state$` property on `MixerStore` now contains a flat object.